ECCO

Documentation for ECCO.jl Julia package.

ECCO stands for Estimating the Circulation and Climate of the Ocean.

ECCO is a 40+ years running effort initiated by Pr. Carl Wunsch at MIT, and which now involves distributed academic projects accross the US (MIT, NASA JPL, UCSD, UTA, WHOI), Europe (AWI, UH), and the UK (BAS).

this package is in early development stage

ECCO.Lorenz_models.L63

ECCO.Lorenz_models.L96

ECCO.Zygote_examples.Zygote_ex1

ECCO.glacier_model.forward_problem

ECCO.toy_problems.Enzyme_ex1

ECCO.toy_problems.Enzyme_ex2

ECCO.toy_problems.Enzyme_ex3

ECCO.toy_problems.Enzyme_ex4

ECCO.toy_problems.ForwardDiff_ex1

ECCO.toy_problems.optim_ex1

ECCO.toy_problems.optim_ex2

ECCO.toy_problems.optim_ex3

ECCO.toy_problems.Enzyme_ex1 — Method

ECCO.toy_problems.Enzyme_ex1()

using ECCO
(f,f_ad,x)=ECCO.toy_problems.Enzyme_ex1()
f(x)
f_ad(x)

ECCO.toy_problems.Enzyme_ex2 — Method

ECCO.toy_problems.Enzyme_ex2()

using ECCO
(f,f_ad,x)=ECCO.toy_problems.Enzyme_ex2()
f(x...)
f_ad(x...)

ECCO.toy_problems.Enzyme_ex3 — Method

ECCO.toy_problems.Enzyme_ex3()

using ECCO
(f,f_ad,x,y)=ECCO.toy_problems.Enzyme_ex3()
f(x,y)
f_ad(x,y)

ECCO.toy_problems.Enzyme_ex4 — Method

ECCO.toy_problems.Enzyme_ex4()

using ECCO
(f,f_ad,x,y)=ECCO.toy_problems.Enzyme_ex4()
f(x,y)
f_ad(x,y)

ECCO.toy_problems.ForwardDiff_ex1 — Method

ECCO.toy_problems.ForwardDiff_ex1()

using ECCO
(x,adx)=ECCO.toy_problems.ForwardDiff_ex1()

ECCO.toy_problems.optim_ex1 — Method

ECCO.toy_problems.optim_ex1()

using ECCO
(f,x0,x1,result)=ECCO.toy_problems.optim_ex1()

ECCO.toy_problems.optim_ex2 — Method

ECCO.toy_problems.optim_ex2()

ECCO.toy_problems.optim_ex3 — Method

ECCO.toy_problems.optim_ex3()

ECCO.Zygote_examples.Zygote_ex1 — Method

ECCO.toy_problems.Zygote_ex1()

using ECCO
(x,adx)=ECCO.Zygote_examples.Zygote_ex1()

ECCO.glacier_model.forward_problem — Function

forward_problem(M0=0.004; dt=1/12.0, nt=6*5000, dx = 1.0, nx = 30)

Simple, 1D mountain glacier model inspired from the book Fundamentals of Glacier Dynamics, by CJ van der Veen, and which was translated to Julia by S Gaikwad.

See https://sicopolis.readthedocs.io/en/latest/AD/tutorial_tapenade.html#mountain-glacier-model

V=ECCO.glacier_model.forward_problem(0.002)

ECCO.Lorenz_models.L63 — Method

L63(; nt=10000)

See https://en.wikipedia.org/wiki/Lorenz_system

using ECCO, CairoMakie
x,y,z=ECCO.Lorenz_models.L63()
lines(x,y)

ECCO.Lorenz_models.L96 — Method

L96(; N=5, F=8)

See https://en.wikipedia.org/wiki/Lorenz96model

using ECCO, CairoMakie
store=ECCO.Lorenz_models.L96()
lines(store[1,:]); lines!(store[2,:]); lines!(store[end,:])
current_figure()